:refactor: Split method, interface and struct definitions
