+++ definitions.go
@@ -47,34 +47,6 @@ type File struct {
 	Files      []*File
 }
 
-type Interface struct {
-	pkg     *Package
-	name    string
-	Methods []MethodDescriptor
-	Doc     Doc
-}
-
-// MethodArgument represent type of fields and arguments.
-type MethodArgument struct {
-	Name string
-	Type RefType
-	Doc  Doc
-}
-
-type MethodDescriptor struct {
-	baseType
-	Doc       Doc
-	Recv      []MethodArgument
-	Arguments []MethodArgument
-	Result    []MethodResult
-	Tag       Tag
-}
-
-type MethodResult struct {
-	Name string
-	Type Type
-}
-
 type Package struct {
 	Name        string
 	ImportPath  string
@@ -361,13 +333,6 @@ func (t *baseType) AddMethod(method *TypeMethod) {
 	t.methods = append(t.methods, method)
 }
 
-type Struct struct {
-	baseType
-	Doc        Doc
-	Fields     []*Field
-	Interfaces []*Interface
-}
-
 type TypeMethod struct {
 	Name       string
 	Descriptor *MethodDescriptor
@@ -402,41 +367,6 @@ func (doc *Doc) FormatComment() string {
 		}
 	}
 	return str
-}
-
-// NewInterface Create new Interface.
-func NewInterface(pkg *Package, name string) *Interface {
-	return &Interface{
-		pkg:  pkg,
-		name: name,
-	}
-}
-
-// Package get name of Pacakge.
-func (i *Interface) Package() *Package {
-	return i.pkg
-}
-
-// Name return name of Struct than implement this Interface.
-func (i *Interface) Name() string {
-	return i.name
-}
-
-// NewMethodDescriptor return the pointer of new MethodDescriptor
-func NewMethodDescriptor(pkg *Package, name string) *MethodDescriptor {
-	return &MethodDescriptor{
-		baseType: *NewBaseType(pkg, name),
-	}
-}
-
-// Package return pointer of Package
-func (method *MethodDescriptor) Package() *Package {
-	return method.pkg
-}
-
-// Name return name of Method
-func (method *MethodDescriptor) Name() string {
-	return method.name
 }
 
 // AppendStruct add new Struct in Package
@@ -514,41 +444,6 @@ func (p *Package) AppendVariable(variable *Variable) *Variable {
 	p.Variables = append(p.Variables, variable)
 	return variable
 }
-
-// NewStruct return new pointer Struct
-func NewStruct(pkg *Package, name string) *Struct {
-	srct := &Struct{
-		baseType: baseType{
-			pkg:  pkg,
-			name: name,
-		},
-	}
-	return srct
-}
-
-// Package return pointer package of Struct
-func (s *Struct) Package() *Package {
-	return s.pkg
-}
-
-// Name return name of Struct
-func (s *Struct) Name() string {
-	return s.name
-}
-
-// FormatComments show struct with all comments
-/*func (s *Struct) FormatComments() string {
-	str := fmt.Sprintf("%s\ntype %s struct {\n", s.Doc.FormatComment(), s.Name())
-	for _, f := range s.Fields {
-		c := f.Doc.FormatComment()
-		brk := "\n"
-		if c == "" {
-			brk = ""
-		}
-		str += fmt.Sprintf("%s%s%s %s %s\n", c, brk, f.Name, f.RefType.Name, f.Tag.Raw)
-	}
-	return fmt.Sprintf("%s}", str)
-}*/
 
 // AppendTagParam add new TagParam in Tag
 func (t *Tag) AppendTagParam(tNew *TagParam) bool {
